+++ .gitignore
@@ -19,6 +19,7 @@
 /src/modules/m_mssql.cpp
 /src/modules/m_mysql.cpp
 /src/modules/m_pgsql.cpp
+/src/modules/m_rand_gcrypt.cpp
 /src/modules/m_regex_pcre.cpp
 /src/modules/m_regex_posix.cpp
 /src/modules/m_regex_re2.cpp
